Bugs fixed.
Amount are now displayed as float, utf8-completely supported.
Transfers now display role and current account code of the dealer.

+++ TIW-exam-pureHTML/src/daos/TransferDAO.java
@@ -29,7 +29,7 @@ public class TransferDAO {
 			while(result.next()) {
 				TransferBean newTransfer = new TransferBean();
 				newTransfer.setIdtransfer(result.getInt("idtransfer"));
-				newTransfer.setAmount(result.getInt("amount"));
+				newTransfer.setAmount(result.getFloat("amount"));
 				newTransfer.setDate(result.getDate("date"));
 				newTransfer.setReason(result.getString("reason"));
 				newTransfer.setCApayer(result.getString("CApayer"));
